+++ owncloudDomain/src/main/java/com/owncloud/android/domain/appregistry/model/AppRegistry.kt
